More size disambiguation

+++ tab/cmd_bighalfwheel.c
@@ -15,6 +15,7 @@
 #include <math.h>
 #include <string.h>
 #include <errno.h>
+#include <stdint.h>
 
 #include "../src/cthugha.h"
 #include "../src/cth_buffer.h"
@@ -28,7 +29,7 @@ int BUFF_HEIGHT;
 
 #define BSIZE (XSIZE * YSIZE)
 
-typedef long tType;
+typedef uint32_t tType;
 #define MS SL
 
 typedef int* trans_map;
